Refactor + fix re-render
